in the process of refactoring to use a global struct

+++ src/long_mode_os/kernel/paging.h
@@ -2,7 +2,7 @@
 
 void paging_init();
 
-void linear_translate_v_to_pmap(int virtual_address,word map[4]);
+void linear_translate_v_to_pmap(long virtual_address,word map[4]);
 
 void map_section();
